ref: move vsts testutils to fixtures
