+++ tests/sentry/integrations/vsts/test_integration.py
@@ -4,6 +4,7 @@ from urllib.parse import parse_qs, urlparse
 import pytest
 import responses
 
+from fixtures.vsts import CREATE_SUBSCRIPTION, VstsIntegrationTestCase
 from sentry.integrations.vsts import VstsIntegration, VstsIntegrationProvider
 from sentry.models import (
     Integration,
@@ -12,8 +13,6 @@ from sentry.models import (
     Repository,
 )
 from sentry.shared_integrations.exceptions import IntegrationError, IntegrationProviderError
-
-from .testutils import CREATE_SUBSCRIPTION, VstsIntegrationTestCase
 
 FULL_SCOPES = ["vso.code", "vso.graph", "vso.serviceendpoint_manage", "vso.work_write"]
 LIMITED_SCOPES = ["vso.graph", "vso.serviceendpoint_manage", "vso.work_write"]
