+++ tests/sentry/integrations/vsts/test_notify_action.py
@@ -3,6 +3,7 @@ from time import time
 import responses
 from freezegun import freeze_time
 
+from fixtures.vsts import GET_PROJECTS_RESPONSE, WORK_ITEM_RESPONSE
 from sentry.integrations.vsts import AzureDevopsCreateTicketAction
 from sentry.integrations.vsts.integration import VstsIntegration
 from sentry.models import ExternalIssue, GroupLink, Identity, IdentityProvider, Integration, Rule
@@ -11,7 +12,6 @@ from sentry.types.rules import RuleFuture
 from sentry.utils import json
 
 from .test_issues import VstsIssueBase
-from .testutils import GET_PROJECTS_RESPONSE, WORK_ITEM_RESPONSE
 
 
 @freeze_time()
